+++ includes/core-phpunit/includes/abstract-testcase.php
@@ -64,6 +64,8 @@ abstract class WP_UnitTestCase_Base extends PHPUnit_Adapter_TestCase {
 
 	/**
 	 * Runs the routine before setting up all tests.
+	 *
+	 * @global wpdb $wpdb WordPress database abstraction object.
 	 */
 	public static function set_up_before_class() {
 		global $wpdb;
@@ -108,8 +110,12 @@ abstract class WP_UnitTestCase_Base extends PHPUnit_Adapter_TestCase {
 
 	/**
 	 * Runs the routine before each test is executed.
+	 *
+	 * @global WP_Rewrite $wp_rewrite WordPress rewrite rules object.
 	 */
 	public function set_up() {
+		global $wp_rewrite;
+
 		set_time_limit( 0 );
 
 		$this->factory = static::factory();
@@ -121,8 +127,6 @@ abstract class WP_UnitTestCase_Base extends PHPUnit_Adapter_TestCase {
 		if ( ! self::$hooks_saved ) {
 			$this->_backup_hooks();
 		}
-
-		global $wp_rewrite;
 
 		$this->clean_up_global_scope();
 
@@ -165,10 +169,17 @@ abstract class WP_UnitTestCase_Base extends PHPUnit_Adapter_TestCase {
 
 	/**
 	 * After a test method runs, resets any state in WordPress the test method might have changed.
+	 *
+	 * @global wpdb     $wpdb         WordPress database abstraction object.
+	 * @global WP_Query $wp_the_query Main WordPress query object.
+	 * @global WP_Query $wp_query     WordPress query object.
+	 * @global WP       $wp           WordPress environment object.
 	 */
 	public function tear_down() {
 		global $wpdb, $wp_the_query, $wp_query, $wp;
+
 		$wpdb->query( 'ROLLBACK' );
+
 		if ( is_multisite() ) {
 			while ( ms_is_switched() ) {
 				restore_current_blog();
@@ -369,10 +380,10 @@ abstract class WP_UnitTestCase_Base extends PHPUnit_Adapter_TestCase {
 	 * Stores $wp_filter, $wp_actions, $wp_filters, and $wp_current_filter
 	 * on a class variable so they can be restored on tear_down() using _restore_hooks().
 	 *
-	 * @global array $wp_filter
-	 * @global array $wp_actions
-	 * @global array $wp_filters
-	 * @global array $wp_current_filter
+	 * @global array $wp_filter         All of the filters and actions.
+	 * @global array $wp_actions        The number of times each action was triggered.
+	 * @global array $wp_filters        The number of times each filter was triggered.
+	 * @global array $wp_current_filter The list of current filters with the current one last.
 	 */
 	protected function _backup_hooks() {
 		self::$hooks_saved['wp_filter'] = array();
@@ -392,10 +403,10 @@ abstract class WP_UnitTestCase_Base extends PHPUnit_Adapter_TestCase {
 	 * Restores the hook-related globals to their state at set_up()
 	 * so that future tests aren't affected by hooks set during this last test.
 	 *
-	 * @global array $wp_filter
-	 * @global array $wp_actions
-	 * @global array $wp_filters
-	 * @global array $wp_current_filter
+	 * @global array $wp_filter         All of the filters and actions.
+	 * @global array $wp_actions        The number of times each action was triggered.
+	 * @global array $wp_filters        The number of times each filter was triggered.
+	 * @global array $wp_current_filter The list of current filters with the current one last.
 	 */
 	protected function _restore_hooks() {
 		if ( isset( self::$hooks_saved['wp_filter'] ) ) {
@@ -417,6 +428,8 @@ abstract class WP_UnitTestCase_Base extends PHPUnit_Adapter_TestCase {
 
 	/**
 	 * Flushes the WordPress object cache.
+	 *
+	 * @global WP_Object_Cache $wp_object_cache WordPress Object Cache object.
 	 */
 	public static function flush_cache() {
 		global $wp_object_cache;
@@ -462,13 +475,15 @@ abstract class WP_UnitTestCase_Base extends PHPUnit_Adapter_TestCase {
 	 *
 	 * @since 5.1.0
 	 *
-	 * @global array $wp_meta_keys
+	 * @global array $wp_meta_keys Global registry for meta keys.
 	 */
 	public function unregister_all_meta_keys() {
 		global $wp_meta_keys;
+
 		if ( ! is_array( $wp_meta_keys ) ) {
 			return;
 		}
+
 		foreach ( $wp_meta_keys as $object_type => $type_keys ) {
 			foreach ( $type_keys as $object_subtype => $subtype_keys ) {
 				foreach ( $subtype_keys as $key => $value ) {
@@ -480,11 +495,15 @@ abstract class WP_UnitTestCase_Base extends PHPUnit_Adapter_TestCase {
 
 	/**
 	 * Starts a database transaction.
+	 *
+	 * @global wpdb $wpdb WordPress database abstraction object.
 	 */
 	public function start_transaction() {
 		global $wpdb;
+
 		$wpdb->query( 'SET autocommit = 0;' );
 		$wpdb->query( 'START TRANSACTION;' );
+
 		add_filter( 'query', array( $this, '_create_temporary_tables' ) );
 		add_filter( 'query', array( $this, '_drop_temporary_tables' ) );
 	}
@@ -493,9 +512,12 @@ abstract class WP_UnitTestCase_Base extends PHPUnit_Adapter_TestCase {
 	 * Commits the queries in a transaction.
 	 *
 	 * @since 4.1.0
+	 *
+	 * @global wpdb $wpdb WordPress database abstraction object.
 	 */
 	public static function commit_transaction() {
 		global $wpdb;
+
 		$wpdb->query( 'COMMIT;' );
 	}
 
@@ -1150,6 +1172,8 @@ abstract class WP_UnitTestCase_Base extends PHPUnit_Adapter_TestCase {
 	 * @since 5.3.0 Formalized the existing `...$prop` parameter by adding it
 	 *              to the function signature.
 	 *
+	 * @global WP_Query $wp_query WordPress Query object.
+	 *
 	 * @param string ...$prop Any number of WP_Query properties that are expected to be true for the current request.
 	 */
 	public function assertQueryTrue( ...$prop ) {
@@ -1649,7 +1673,7 @@ abstract class WP_UnitTestCase_Base extends PHPUnit_Adapter_TestCase {
 	 *
 	 * @since 4.4.0
 	 *
-	 * @global WP_Rewrite $wp_rewrite
+	 * @global WP_Rewrite $wp_rewrite WordPress rewrite rules object.
 	 *
 	 * @param string $structure Optional. Permalink structure to set. Default empty.
 	 */